Changed the code from ROP to majorly OOP
