+++ package/GVA/dns_recon.py
@@ -1,156 +1,72 @@
-import json
-import re
-from typing import Any
-from typing import Optional
-
-import dns.resolver
-import openai
+from typing import Any, Optional
+import requests
+import dns.resolver as dns_resolver_module
 from rich.progress import track
 
-model_engine = "text-davinci-003"
 
-
-def extract_data(json_string: str) -> Any:
-    # Define the regular expression patterns for individual values
-    A_pattern = r'"A": \["(.*?)"\]'
-    AAA_pattern = r'"AAA: \["(.*?)"\]'
-    NS_pattern = r'"NS": \["(.*?)"\]'
-    MX_pattern = r'"MX": \["(.*?)"\]'
-    PTR_pattern = r'"PTR": \["(.*?)"\]'
-    SOA_pattern = r'"SOA": \["(.*?)"\]'
-    TXT_pattern = r'"TXT": \["(.*?)"\]'
-
-    # Initialize variables for extracted data
-    A = None
-    AAA = None
-    NS = None
-    MX = None
-    PTR = None
-    SOA = None
-    TXT = None
-
-    # Extract individual values using patterns
-    match = re.search(A_pattern, json_string)
-    if match:
-        A = match.group(1)
-
-    match = re.search(AAA_pattern, json_string)
-    if match:
-        AAA = match.group(1)
-
-    match = re.search(NS_pattern, json_string)
-    if match:
-        NS = match.group(1)
-
-    match = re.search(MX_pattern, json_string)
-    if match:
-        MX = match.group(1)
-
-    match = re.search(PTR_pattern, json_string)
-    if match:
-        PTR = match.group(1)
-
-    match = re.search(SOA_pattern, json_string)
-    if match:
-        SOA = match.group(1)
-
-    match = re.search(TXT_pattern, json_string)
-    if match:
-        TXT = match.group(1)
-
-    # Create a dictionary to store the extracted data
-    data = {
-        "A": A,
-        "AAA": AAA,
-        "NS": NS,
-        "MX": MX,
-        "PTR": PTR,
-        "SOA": SOA,
-        "TXT": TXT
-    }
-
-    # Convert the dictionary to JSON format
-    json_output = json.dumps(data)
-
-    return json_output
-
-
-def AI(analyze: str, key: Optional[str]) -> str:
-    openai.api_key = key
-    prompt = f"""
-    Do a DNS analysis on the provided DNS scan information
-    The DNS output must return in a JSON format accorging to the provided
-    output format. The data must be accurate in regards towards a pentest report.
-    The data must follow the following rules:
-    1) The DNS scans must be done from a pentester point of view
-    2) The final output must be minimal according to the format given
-    3) The final output must be kept to a minimal
-
-    The output format:
-    {{
-        "A": [""],
-        "AAA": [""],
-        "NS": [""],
-        "MX": [""],
-        "PTR": [""],
-        "SOA": [""],
-        "TXT": [""]
-    }}
-
-    DNS Data to be analyzed: {analyze}
-    """
-    try:
-        # A structure for the request
-        completion = openai.Completion.create(
-            engine=model_engine,
-            prompt=prompt,
-            max_tokens=1024,
-            n=1,
-            stop=None,
-        )
-        response = completion.choices[0].text
-        return extract_data(str(response))
-    except KeyboardInterrupt:
-        print("Bye")
-        quit()
-
-
-def dns_recon(target: Optional[str], key: str) -> str:
-    if key is not None:
-        pass
-    else:
-        raise ValueError("KeyNotFound: Key Not Provided")
-    if target is not None:
-        pass
-    else:
-        raise ValueError("InvalidTarget: Target Not Provided")
-    analyze = ''
-    # The DNS Records to be enumeratee
-    record_types = ['A', 'AAAA', 'NS', 'CNAME', 'MX', 'PTR', 'SOA', 'TXT']
-    for records in track(record_types):
-        try:
-            answer = dns.resolver.resolve(target, records)
-            for server in answer:
-                st = server.to_text()
-                analyze += "\n"
-                analyze += records
-                analyze += " : "
-                analyze += st
-        except dns.resolver.NoAnswer:
-            print('No record Found')
+class DNSRecon:
+    def dns_resolver(self, AIModels, target: str, akey: Optional[str], bkey: Optional[str], lkey, lendpoint, AI: str) -> Any:
+        if target is not None:
             pass
-        except dns.resolver.NXDOMAIN:
-            print('NXDOMAIN record NOT Found')
-            pass
-        except dns.resolver.LifetimeTimeout:
-            print("Timmed out check your internet")
-            pass
-        except KeyboardInterrupt:
-            print("Bye")
-            quit()
-    try:
-        response = AI(key, analyze)
+        else:
+            raise ValueError("InvalidTarget: Target Not Provided")
+        analyze = ''
+        # The DNS Records to be enumerated
+        record_types = ['A', 'AAAA', 'NS', 'CNAME', 'MX', 'PTR', 'SOA', 'TXT']
+        for record_type in track(record_types):
+            try:
+                answer = dns_resolver_module.resolve(target, record_type)
+                for server in answer:
+                    st = server.to_text()
+                    analyze += f"\n{record_type} : {st}"
+            except dns_resolver_module.NoAnswer:
+                print('No record Found')
+                pass
+            except dns_resolver_module.NXDOMAIN:
+                print('NXDOMAIN record NOT Found')
+                pass
+            except dns_resolver_module.LifetimeTimeout:
+                print("Timed out, check your internet")
+                pass
+            except requests.exceptions.InvalidHeader:
+                pass
+            except KeyboardInterrupt:
+                print("Bye")
+                quit()
+
+        response = ""
+        match AI:
+            case 'openai':
+                try:
+                    if akey is not None:
+                        # Clean up Bearer token from newline characters
+                        akey = akey.replace('\n', '')
+                    else:
+                        raise ValueError("KeyNotFound: Key Not Provided")
+                    response = AIModels.gpt_ai(akey, analyze)
+                except KeyboardInterrupt:
+                    print("Bye")
+                    quit()
+            case 'bard':
+                try:
+                    if bkey is not None:
+                        bkey = bkey.replace('\n', '')
+                    else:
+                        raise ValueError("KeyNotFound: Key Not Provided")
+                    response = AIModels.BardAI(bkey, analyze)
+                except KeyboardInterrupt:
+                    print("Bye")
+                    quit()
+            case 'llama':
+                try:
+                    response = AIModels.llama_AI(analyze, "local", lkey, lendpoint)
+                except KeyboardInterrupt:
+                    print("Bye")
+                    quit()
+            case 'llama-api':
+                try:
+                    response = AIModels.llama_AI(analyze, "runpod", lkey, lendpoint)
+                except KeyboardInterrupt:
+                    print("Bye")
+                    quit()
         return str(response)
-    except KeyboardInterrupt:
-        print("Bye")
-        quit()
