+++ package/setup.cfg
@@ -1,11 +1,9 @@
 [metadata]
 name = GVA
-version = 0.1.1
-version = 0.1.1
+version = 1.1.8
 author = Chiranjeevi G
 author_email = morpheuslord@protonmail.com
-description = ChatGPT vuln analysis with python
-description = ChatGPT vuln analysis with python
+description = Vuln analyzer and other basic recon tools with AI capabilities
 long_description = file: README.md
 long_description_content_type = text/markdown
 url = https://github.com/morpheuslord/GVA_package
