+++ package/setup.py
@@ -9,8 +9,7 @@ here = os.path.abspath(os.path.dirname(__file__))
 with codecs.open(os.path.join(here, "README.md"), encoding="utf-8") as fh:
     long_description = "\n" + fh.read()
 
-VERSION = '0.1.1'
-VERSION = '0.1.1'
+VERSION = '1.1.8'
 DESCRIPTION = 'Python Project for GPT-Vuln_analyzer'
 LONG_DESCRIPTION = """
     This is a Proof Of Concept application that demostrates how AI can be used to generate accurate
